+++ CHANGELOG.md
@@ -10,6 +10,12 @@ with respect to its command line interface and HTTP interface.
 ### Fixed
 * CLI + Server: No longer panics when printing a nil Deployment.
 
+### Fixed
+* Server: Profiling configuration - accepts SOUS_PROFILING environment variable correctly.
+
+### Changed
+* Developer: Server action extracted.
+
 ## [0.5.56](//github.com/opentable/sous/compare/0.5.55...0.5.56)
 ### Fixed
 * CLI: No longer panics under normal operation (e.g. when trying to run 'sous deploy'